remove unnecessary module loading
