+++ utilities/createnc.py
@@ -18,6 +18,79 @@ import datetime
 import argparse
 
 
+def get_state_interpolator(state, field=0):
+    """
+    Get a Scipy interpolation object for a field on a AMR grid.
+    """
+    import scipy.interpolate
+
+    # the underlying patch in this state object
+    p = state.patch
+
+    # x, y arrays and also dx, dy for checking
+    x, dx = numpy.linspace(p.lower_global[0]+p.delta[0]/2.,
+                           p.upper_global[0]-p.delta[0]/2.,
+                           p.num_cells_global[0], retstep=True)
+    y, dy = numpy.linspace(p.lower_global[1]+p.delta[1]/2.,
+                           p.upper_global[1]-p.delta[1]/2.,
+                           p.num_cells_global[1], retstep=True)
+    assert numpy.abs(dx-p.delta[0]) < 1e-6, "{} {}".format(dx, p.delta[0])
+    assert numpy.abs(dy-p.delta[1]) < 1e-6, "{} {}".format(dy, p.delta[1])
+
+    # get the interpolation object
+    kx = ky = 3
+
+    if x.size <= 3:
+        kx = x.size - 1
+
+    if y.size <= 3:
+        ky = y.size - 1
+
+    interp = scipy.interpolate.RectBivariateSpline(
+        x, y, state.q[field, :, :],
+        [p.lower_global[0], p.upper_global[0], p.lower_global[1], p.upper_global[1]],
+        kx=kx, ky=ky)
+
+    return interp
+
+def interpolate(solution, x_target, y_target,
+                field=0, shift=[0., 0.], level=1,
+                clip=True, clip_less=1e-7, nodatavalue=-9999.):
+    """
+    Do the interpolation.
+    """
+
+    # allocate space for interpolated results
+    values = numpy.zeros((y_target.size, x_target.size), dtype=numpy.float64)
+
+    # loop through all AMR grids
+    for state in solution.states:
+
+        p = state.patch
+
+        # only do subsequent jobs if this is at the target level
+        if p.level != level:
+            continue
+
+        # get the indices of the target coordinates that are inside this patch
+        xid = numpy.where((x_target>=p.lower_global[0])&(x_target<=p.upper_global[0]))[0]
+        yid = numpy.where((y_target>=p.lower_global[1])&(y_target<=p.upper_global[1]))[0]
+
+        # get interpolation object
+        interpolator = get_state_interpolator(state, field)
+
+        # if any target coordinate located in thie patch, do interpolation
+        if xid.size and yid.size:
+            values[yid[:, None], xid[None, :]] = \
+                interpolator(x_target[xid]-shift[0], y_target[yid]-shift[1]).T
+
+    # apply nodatavalue to a threshold
+    if clip:
+        values[values<clip_less] = nodatavalue
+
+    return values
+
+
 if __name__ == "__main__":
 
     # get the abs path of the repo
@@ -33,7 +106,7 @@ if __name__ == "__main__":
     sys.path.insert(0, claw_dir)
 
     from clawpack import pyclaw
-    from pphelper import get_bounding_box, interpolate
+    from pphelper import get_bounding_box
 
     # CMD argument parser
     parser = argparse.ArgumentParser(
